Add support for SF2e
