+++ monks-tokenbar.js
@@ -21,6 +21,7 @@ import { SwadeRolls } from "./systems/swade-rolls.js";
 import { SW5eRolls } from "./systems/sw5e-rolls.js";
 import { CoC7Rolls } from "./systems/coc7-rolls.js";
 import { T2K4ERolls } from "./systems/t2k4e-rolls.js";
+import { SF2eRolls } from "./systems/sf2e-rolls.js";
 
 
 export let debug = (...args) => {
@@ -173,7 +174,7 @@ export class MonksTokenBar {
                 let token = fromUuidSync(msgToken.uuid);
                 let actor = token?.actor ? token.actor : token;
 
-                return game.system.id == "pf2e" && !!msgToken.roll && actor.type == "character" && actor.heroPoints.value > 0 && (game.user.isGM || actor.isOwner);
+                return ["pf2e", "sf2e"].includes(game.system.id) && !!msgToken.roll && actor.type == "character" && actor.heroPoints.value > 0 && (game.user.isGM || actor.isOwner);
 
                 //return game.user.isGM && !!message.flags.pf2e.context
                 //message.isRerollable && !!actor?.isOfType("character") && actor.heroPoints.value > 0
@@ -1193,7 +1194,7 @@ export class MonksTokenBar {
         let requesttype = a.dataset.requesttype.toLowerCase();
         if (requesttype == 'request')
             MonksTokenBarAPI.requestRoll(canvas.tokens.controlled, options);
-        else if (requesttype == 'contested') 
+        else if (requesttype == 'contested')
             MonksTokenBarAPI.requestContestedRoll({ request: options.request }, { request: options.request1 }, options);
     }
 }
@@ -1243,6 +1244,8 @@ Hooks.on("setup", () => {
             MonksTokenBar.system = new PF1Rolls(); break;
         case 'pf2e':
             MonksTokenBar.system = new PF2eRolls(); break;
+        case 'sf2e':
+            MonksTokenBar.system = new SF2eRolls(); break;
         case 'tormenta20':
             MonksTokenBar.system = new Tormenta20Rolls(); break;
         case 'sfrpg':
@@ -1254,7 +1257,7 @@ Hooks.on("setup", () => {
         case 'coc7':
             MonksTokenBar.system = new CoC7Rolls(); break;
         case 't2k4e':
-            MonksTokenBar.system = new T2K4ERolls(); break; 
+            MonksTokenBar.system = new T2K4ERolls(); break;
     }
 
     MonksTokenBar.system.constructor.activateHooks();
