+++ systems/sf2e-rolls.js
@@ -1,0 +1,3 @@
+import { getRollsForPf2eBasedSystem } from "./pf2e-base.js";
+
+export const SF2eRolls = getRollsForPf2eBasedSystem('sf2e');
